maint: clarify that .x-update-copyright entries are EREs
The entries in the exemption list are processed by
"grep -vEf ./.x-update-copyright" and therefore evaluated as an
extended regular expression (ERE).  Thus, e,g, a "bootstrap" entry
would also match for "bootstrap.conf" which we may want to be updated.
The current entries are unambiguous so far, but to avoid confusion when
adding further entries in the future, turn them into proper EREs.

* .x-update-copyright: Change all entries to EREs, i.e. including
the caret ^ and dollar sign $ meta-characters matching the beginning
and the end of a line.